Added context for `UserContext`
