+++ app/components/LikeImageButton/LikeImageButton.tsx
@@ -1,12 +1,13 @@
+import React from "react";
 import { HeartFilled, HeartOutlined, HeartTwoTone } from "@ant-design/icons";
 import { Button } from "antd";
 import type { ImageType } from "~/types";
 import { useRemixFetcher } from "~/hooks";
-import { useLoaderData } from "@remix-run/react";
+import { UserContext } from "~/context";
 
 const LikeImageButton = ({ imageData }: { imageData: ImageType }) => {
-  const loaderData = useLoaderData();
-  const userId = loaderData.user?.id || undefined;
+  const userData = React.useContext(UserContext);
+  const userId = userData?.id || undefined;
   const { fetcher, isLoadingFetcher } = useRemixFetcher();
 
   const handleLikeImage = () => {
